DSA IN C++
Binary Search is another method of searching an element in O(logn) which is the best case complexity. In this method a problem is to be divide in two sub-parts. And then user enter a key which is to be find in an given array.
Process:- First we will find a mid element in an array.
mid= (start+(high-start))/2
(A) If mid element is equal to key then return index or mid.
(B) If array of mid is greater than key then we will move in the left most of an array.
 and end will be mid-1.
(C) If array of mid is smaller than key then we will move in the right most of an array.
 and end will be mid+1.
(D) At last, mid is to be updated.