Create Reverse an array
In this question, if we look the question on coding ninja then there is one tromma which is there is given index value after that we have to reverse an array. Sometimes we get confused after read that question but concept is same only we have to initialise start symbol with 'm+1' and end symbol with 'arr.size()-1'.